Refactoring and fix test flakiness
